+++ SubCodes/UnKronPolicyIndexes/UnKronPolicyIndexes3_FHorz_noz.m
@@ -18,16 +18,16 @@ divisors2=cumprod([1,n_daprime2(1:end-1)])';   % [l_daprime2,1]
 divisors3=cumprod([1,n_daprime3(1:end-1)])';   % [l_daprime3,1]
 
 if vfoptions.gridinterplayer==1
-    Policy=[mod(floor(double(PolicyKron(1,:)-1)./divisors1),n_daprime1(:))+1;
-            mod(floor(double(PolicyKron(2,:)-1)./divisors2),n_daprime2(:))+1;
-            mod(floor(double(PolicyKron(3,:)-1)./divisors3),n_daprime3(:))+1;
+    Policy=[mod(floor((PolicyKron(1,:)-1)./divisors1),n_daprime1(:))+1;
+            mod(floor((PolicyKron(2,:)-1)./divisors2),n_daprime2(:))+1;
+            mod(floor((PolicyKron(3,:)-1)./divisors3),n_daprime3(:))+1;
             PolicyKron(4,:);
             PolicyKron(5,:)];
     Policy=reshape(Policy,[l_daprime1+l_daprime2+l_daprime3+2,n_a,N_j]);
 else
-    Policy=[mod(floor(double(PolicyKron(1,:)-1)./divisors1),n_daprime1(:))+1;
-            mod(floor(double(PolicyKron(2,:)-1)./divisors2),n_daprime2(:))+1;
-            mod(floor(double(PolicyKron(3,:)-1)./divisors3),n_daprime3(:))+1];
+    Policy=[mod(floor((PolicyKron(1,:)-1)./divisors1),n_daprime1(:))+1;
+            mod(floor((PolicyKron(2,:)-1)./divisors2),n_daprime2(:))+1;
+            mod(floor((PolicyKron(3,:)-1)./divisors3),n_daprime3(:))+1];
     Policy=reshape(Policy,[l_daprime1+l_daprime2+l_daprime3,n_a,N_j]);
 end
