+++ ValueFnIter/FHorz/OnCPU/ValueFnIter_FHorz_Par1_raw.m
@@ -85,6 +85,6 @@ end
 %%
 Policy2=zeros(2,N_a,N_z,N_j); %NOTE: this is not actually in Kron form
 Policy2(1,:,:,:)=shiftdim(rem(Policy-1,N_d)+1,-1);
-Policy2(2,:,:,:)=shiftdim(ceil(double(Policy)/N_d),-1);
+Policy2(2,:,:,:)=shiftdim(ceil((Policy)/N_d),-1);
 
 end
